requirements.txt
streamlit
fastapi
uvicorn
pydantic
langchain-community
langchain-groq
pandas
python-dotenv
python-multipart
sqlalchemy
